requirements.txt
streamlit>=1.18.1
pandas>=1.5.3
openpyxl>=3.0.10        # needed by pandas.read_excel
Pillow>=9.5.0           # for Image.open